Link opflex agent with Static boost.
agent-ovs is compiled and tested with 1.58 version of boost.

Some redhat distros do not have this older version of boost,
and installing agent-ovs rpm on such systems can cause issues.

So we link the boost libraries statically and eliminate this
dependency.

Since this involves linking against other shared libraries, it
requires that libboost*.a objects be compiled with -fPIC option.

This can be done on any build system once and copied to
/usr/local/boost_1_58_0 as follows

    wget http://sourceforge.net/projects/boost/files/boost/1.58.0/boost_1_58_0.tar.gz
    tar zxvf boost_1_58_0.tar.gz
    cd boost_1_58_0/
    ./bootstrap.sh --prefix=/usr/local/boost_1_58_0
    ./b2 cxxflags=-fPIC cflags=-fPIC -a
    ./b2 install --prefix=/usr/local/boost_1_58_0

The rpmbuild would copyin this folder to /usr/local within chroot

The agent and libopflex code is built as follows.

    ./autogen.sh
    ./configure --with-static-boost --with-boost=/usr/local/boost_1_58_0

This would include the .a(s) inside the artifcats.

Note that this option is *DISABLED* by default. So if not specified it should
not affect existing builds.

Signed-off-by: Madhu Challa <challa@gmail.com>