Reset learned peers only when server triggers fallback (#349)
Existing peers are closed asynchronously so disconnects
might be processed after the call to readd the configured
peers. In some cases this resulted in an agent with no peers.
Restarting the agent was the only way to recover

Changing to disconnect the learned peers only. Configured
peers do not need to be disconnected

Signed-off-by: Tom Flynn <tom.flynn@gmail.com>