Add Flow support for SRV records
With an application that supports headless services,
a DNS name can resolve to hostname:proto:port and
in this case, currently prefer port/proto from DNS policy
over classifier.
TODO: If required a knob can be added
to choose the behavior in this case.But with DNS policy other
fields should not be set normally.

Signed-off-by: Kiran Shastri <shastrinator@gmail.com>