Revert crash fix for kmr2
Bosch wants no changes and this was a fix added along with async json parse,
the feature itself is enabled under a config option, except the fix for
the double free. So reverting this fix which if hit will cause opflex
to crash and restart.

This revert is specific to this branch and master still has this fix.

Signed-off-by: Madhu Challa <challa@gmail.com>