Add support for multiple l3outs in same VRF
When NAT is used with multiple l3outs, the route table flow
 7. ip,reg6=0x2, priority 80
    load:0x80000003->NXM_NX_REG2[]
    load:0x7b8019->NXM_NX_REG7[]
    write_metadata:0x2/0xff
uses the EPG, REG7 based on VRF match reg6 for traffic from EP to
outside world. This causes an issue since we cannot let the EP pick
which l3out to use.

For revnat its not an issue since there are 2 flows one for each l3out
 7. ip,reg0=0x7a8014,reg6=0x1,nw_dst=60.60.60.184, priority 452
    set_field:00:22:bd:f8:19:ff->eth_src
    set_field:fa:16:3e:9f:60:ba->eth_dst
    set_field:10.10.10.234->ip_dst
    dec_ttl
    load:0x720013->NXM_NX_REG2[]
    load:0x1->NXM_NX_REG4[]
    load:0x1->NXM_NX_REG5[]
    load:0x2->NXM_NX_REG6[]
    load:0xe->NXM_NX_REG7[]
    write_metadata:0x400/0x400

The reason for the inconsistency is that both flows are added as part of external
rd update. The first flow should be based on an ep update since the ep knows
which l3out to use based on its ipm.

Added a new table pre route, updated via ep update to pick preferred nat epg
based on ipm in ep file. The route table then matches on the l3out and updates
external network id.

Signed-off-by: Madhu Challa <challa@gmail.com>