Make policy retry delay configurable
The retry delay should be configurable, with a default initial
backoff value of 10 seconds, and a minimum value of 1 second
(the processor can actually do milliseconds, but we don't see
a need to support this). This can be set to larger amounts
for servers that crash under heavy loads from frequent re-requests.

The patch also introduces an internal class for creating randomized
deltas, which are used to provide some dithering for retries.

(cherry picked from commit e5d0b8959adac24d4208179f65bc5f657015909f)