Fix pre-handshake policy resolve refresh timeouts
If an MO reference is added locally, the refresh timeout for that MO
is in most cases set to half the PRR, or 1 hour with the default timeout
values. If the Processor invokes the OpflexPool's sendToRole handler
before the peer has completed the identity request handshake, then the
OpflexPool won't send the message, and the MO's timeout is unchanged.
When the handshake does happen, the processor re-sends the policy
resolve requests for those MOs, but it doesn't update the timeouts
for those MOs. This can lead to excessive timeout values in some cases
(e.g. request gets rejected with an error, even if the error state is
short-lived).

This patch adds updates of the refresh time for the handshake
succeeded callback, in order to avoid the excessive timeout
described above.