Fix broken shared_ptr compares
These result in updateEPGDomains always returning true
that significantly increases the amount of time updateDomain
takes when a large number of subnets are present.

With my config the order of improvement is from 30 mins down
to 4-5 mins.

To further improve performance a second enhancement would be
to defer the common part than run on every update.

Signed-off-by: Madhu Challa <challa@gmail.com>