Add configurable timeouts for mcast and switch sync
Mcast -
On startup agent will cache old values if present
in opflex-groups.json and start a timer to timeout
these values. Until the timer expires new values
learned are merged with the old values and added
to the file. Eventually the timer will expire and
new values take effect from then on.
        "timers": {
            "mcast-cache-timeout": 600,
        }
section in opflex-agent.conf controls the timeout
in seconds. Default is 300 secs or 5 mins

Switch-state-sync -
        "timers": {
            "switch-sync-delay": 300
        }
This value was previous hardcoded to 5 seconds,
which is also the default. However it can be
changed via the above config option. It needs
to be increased if we anticipate many flows that
exist and a long delay between getting PlatformConfig
Mo and the rest of the Mos that would be needed
to configure the new flowmods. On a fresh system
the value can be 0. (I will see if it can be dynamically
adjusted based on the current flow mods)

Signed-off-by: Madhu Challa <challa@gmail.com>