Skip reset.conf if we processed PlatformConfig del
There are 2 scenarios
1. reset.conf comes immediately after PlatformConfig delete
We will skip processing it if within a certain time, 5 secs default
This is so we do not trigger a disconnect twice for the same event,
one via PlatformConfig delete and other via reset.conf

2. PlatformConfig delete comes after reset.conf
This should not be an issue since we will not be able to
get the PlatformConfig event once the connection has been
reset via reset.conf processing. There is a small race here
and in that worst case we may process the event twice but
the window is significantly reduced.

Signed-off-by: Madhu Challa <challa@gmail.com>