Reset all peers during vmotion interval
Use outOfBandConfigUpdated and outOfBandConfigDeleted to figure
out interval and during this period if a disconnect happens
we will reset both peers and fallback to configured list

This avoids issues where we miss the platformConfigDelete
event during vmotion. This basically does the same thing
as what the reset.conf fix would do in case we miss that
event.

Signed-off-by: Madhu Challa <challa@gmail.com>