Add support for connection tracking nat
1. persistant dnat using CT
This commit adds support to do connection tracking nat per service
by defining a new filed in the service file "conntrack-nat": true
default is false.

When enabled we use conntrack to do the nat for us. As a result
some fine grained flows per next hop will not be there for the
reverse flows. So we lose the capability to do stats, at least
some of the stats.

loopback "-trk" flows were conflicting ctnat flows, hence removed
ct from loopback. loopback is between service backend <> service
and it does not really make sense to ct those.

2. when next-hop moves from next-hop-ips[] to terminating-next-hop-ips[]
allow +est flows for that next-hop but dont allow +new flows.
table=5 is adjusted so new flows will not get hashed to the next-hop
that moved.

Signed-off-by: Madhu Challa <challa@gmail.com>