Add support for connection tracking nat
This commit adds support to do connection tracking nat per service
by defining a new filed in the service file "conntrack-nat": true
default is false.

When enabled we use conntrack to do the nat for us. As a result
some fine grained flows per next hop will not be there for the
reverse flows. So we lose the capability to do stats, at least
some of the stats.

Signed-off-by: Madhu Challa <challa@gmail.com>