Add TCP connection watchdog timer
The agent has hit corner cases where it fails to complete the TCP
connection process with the anycast address. This patch adds a watchdog
timer to the Peer, ensuring that the Peer isn't left in an unconnected
TCP state for more than the configured timeout value (default is value
of 10 seconds, with a value of 0 to disable the feature).

(cherry picked from commit 968ebb7a21c27305291fe9378769a4ae3b6fc5b7)