Introduce Event Notification Manager
A new mechanism is introduced for generating events. The existing
socket-based notifications are enhanced to include file-based
notifications. The configuration of socket vs. file based notifies
is covered by the addition of a notification "mode" field in the
"notif" section, along with the path:

        "notif": {
            "enabled": true,
            "mode": "file",
            "events-dir": "/var/lib/opflex-agent-ovs/events"
        }

The agent will look for files in this directory with .subscription
suffixes, and when present, will parse them to look for event
notification requests. The subscriptions file will use JSON
syntax to describe the events for notifications. For example:

{
    "uuid": "2f5cd01d-7a7c-4156-b2cd-42565b33a52d",
    "time-zone": "UTC",
    "time-format": "UNUIX",
    "subscriptions": [
        {
            "type": "class",
            "state": "deleted",
            "subject": "PlatformConfig"
        }
    ]
}

This requests that the PlatformConfig MO's "deleted" event
is subscribed to. Supported types are "uri" or "class", for
either class-based or instance-based notifications. The state
can be "created", "updated", "deleted", or "any". The subject
is the MO class requested. An optional "uri" field must be
provided for instance-based subscriptions.

File-based notifications are profided in .notifications files
that have matching prefixes (e.g. "foo.subscriptions" would
generate a "foo.notifictions" file). The file is JSON formatted,
with content like the following:

{
    "uuid": "unique-subscription-id",
    "events": [
        {
            "uri": "PolicyUniverse/PlatformConfig/example-config/",
            "timestamp": "2024-01-15 10:30:45.123",
            "state": "deleted"
        }
    ]
}

The timestamp field uses the time zone and format specified
in the .subscriptions file. The time zones supported are "UTC",
and "localtime" (default). The time formats supported include
"ISO8601", "RFC3339", and "UNIX".

The notifidation files are not historical - they are always
overwrititen using the most recently generated event.

The manager currently only supports the PlatformConfig class,
with the "deleted" state, but could be extended to support any
MO. A future patch should add support for the same subscription
mechanism and notifications via the notification socket. A future
patch should also include an interface class, which clients would
implement as callbacks, along with regisgtering with the manager
class to request those events.

(cherry picked from commit 35bfac51986cbd167f8dfe9a64fafb7802b250e3)