CSCwr85402 Ignore frag filter
If "fragmentFlags" is present in a rule simply ignore them.

The fragmentFlags are supposed to match on second packet on that
has a fragment bit set. The issue with this match is that since its
a fragment it cannot match on ports so it effectively becomes an
ip allow all filter without regard to ports that overrides any other
filter with specific ports.

Its assumed that a frag filter will be applied as an independent filter
and not merged with any port specific filter.

If CT is also enabled as such with a port match, Ovs will automatically
reassemble / refragment all the packets for that port match to work. So
there is no specific need of a fragment filter.

Of course if the frag filter is the only allow rule configured that also
will not be supported.