Include reset of configured peers
The EndpointManager gets a callback when the PlatformConfig MO has
been deleted, which causes it to call the OFFramework's method to
resetAllUnconfiguredPeers, which disconnects and closes the connections
to any unconfigured peers, and then adds the configured peers to the
the list of peers to connect to. There is a race condition where the
agent may not yet have closed the connection of a configured peer when
this call happens, resulting in a sequence where the agent skips adding
the configured peer, and then the configured peer connection closes,
leaving the agent in a state where it's not connected to any peers and
doesn't think it has any peers to connect to.

To avoid this, we add a call to reset all peers, including any
configured ones, in order to avoid this state.