Add opflex connection monitoring script
The opflex agent has encountered corner cases during switch reboots,
which cause the agent to stop attempting to reconnect to the fabric.
This patch adds a script that monitors for connections to the opflex
TCP port, and if no connections are detected for a continued period
of time, it writes to the reset.conf file, triggering the reconnect
process.

(cherry picked from commit f79bd8cf0aa0f5fd48e65e0f221582a5555c7c66)